Done, without tests

+++ src/main/java/ee/ivkhkdev/carpartshop/model/Product.java
@@ -2,4 +2,5 @@ package ee.ivkhkdev.carpartshop.model;
 
 import java.io.Serializable;
 
-public record Product(String name, double price) implements Serializable {}
+public record Product(String name, float price) implements Serializable { }
+
